[WIP] move the cpu constructor -> gpu constructor pass to joint graph
[ghstack-poisoned]
