+++ torch/_inductor/fx_passes/joint_graph.py
@@ -10,6 +10,10 @@ import torch._guards
 import torch.utils._pytree as pytree
 from torch._inductor.constant_folding import ConstantFolder
 from torch._inductor.fx_passes.dedupe_symint_uses import _SymHashingDict
+from torch._inductor.fx_passes.move_constructors_to_device import (
+    move_constructors_to_gpu,
+)
+from ..fx_utils import FakeTensorUpdater
 from torch.fx.experimental.symbolic_shapes import statically_known_true
 from torch.fx.passes.graph_transform_observer import GraphTransformObserver
 from torch.multiprocessing.reductions import StorageWeakRef
@@ -462,6 +466,12 @@ def joint_graph_passes(graph: torch.fx.GraphModule):
     if not config.fallback_random:
         count += replace_random_passes(graph)
 
+    fake_tensor_updater = FakeTensorUpdater(graph.graph)
+
+    move_constructors_to_gpu(graph.graph)
+
+    fake_tensor_updater.incremental_update()
+
     if config.joint_custom_post_pass is not None:
         with GraphTransformObserver(
             graph, "joint_custom_post_pass", config.trace.log_url_for_graph_xform
@@ -474,59 +484,6 @@ def joint_graph_passes(graph: torch.fx.GraphModule):
         graph.graph.lint()
         graph.recompile()
     return graph
-
-
-@register_graph_pattern(
-    CallFunction(
-        torch.ops.prims.iota.default,
-        KeywordArg("length"),
-        start=KeywordArg("start"),
-        step=KeywordArg("step"),
-        dtype=KeywordArg("dtype"),
-        device=KeywordArg("device"),
-        requires_grad=KeywordArg("requires_grad"),
-    ),
-    pass_dict=patterns,
-)
-def fix_iota_device(match: Match, length, start, step, dtype, device, requires_grad):
-    """
-    Eager supports:
-
-        aten.index(cuda_tensor, torch.arange(..., device="cpu"))
-
-    But this results in an implicit host-device-copy and breaks cudagraphs.
-    Rewrite the arange to use CUDA.
-    """
-    (node,) = match.nodes
-    user_devices: OrderedSet[torch.device] = OrderedSet()
-    for user in node.users:
-        if (
-            user.op == "call_function"
-            and user.target in (aten.index.Tensor, aten.index_put.default)
-            and hasattr(user.meta.get("val"), "device")
-        ):
-            user_devices.add(user.meta["val"].device)  # type: ignore[union-attr]
-        else:
-            return  # bail out
-
-    if len(user_devices) == 1 and "val" in node.meta:
-        (user_device,) = user_devices
-        if device.type != user_device.type:
-            repl = match.graph.call_function(
-                torch.ops.prims.iota.default,
-                (length,),
-                {
-                    "start": start,
-                    "step": step,
-                    "dtype": dtype,
-                    "device": user_device,
-                    "requires_grad": requires_grad,
-                },
-            )
-            repl.meta.update(node.meta)
-            repl.meta["val"] = repl.meta["val"].to(user_device)
-            node.replace_all_uses_with(repl)
-            match.erase_nodes()
 
 
 @register_graph_pattern(
