+++ torch/_inductor/fx_passes/post_grad.py
@@ -4,8 +4,7 @@ import functools
 import itertools
 import logging
 import operator
-from collections import Counter, defaultdict
-from typing import Any, Dict, List, Optional, Set, TYPE_CHECKING, Union
+from typing import Any, Dict, List, Optional, TYPE_CHECKING, Union
 
 import torch
 import torch._inductor as inductor
@@ -42,7 +41,7 @@ from ..pattern_matcher import (
     register_graph_pattern,
     stable_topological_sort,
 )
-from ..utils import decode_device, get_gpu_type, is_pointwise_use
+from ..utils import decode_device, is_pointwise_use
 from ..virtualized import V
 from .b2b_gemm import B2B_GEMM_PASS
 from .ddp_fusion import fuse_ddp_communication
@@ -131,8 +130,6 @@ def post_grad_passes(gm: torch.fx.GraphModule, is_inference: bool):
             config.post_grad_custom_post_pass(gm.graph)
 
     stable_topological_sort(gm.graph)
-
-    move_constructors_to_gpu(gm.graph)
 
     fake_tensor_updater.incremental_update()
 
@@ -1083,236 +1080,3 @@ def check_shape_cuda_and_fused_int_mm_mul_enabled(match):
 )
 def fused_int_mm_mul(match: Match, mat1, mat2, mat3, out_dtype=None):
     return inductor.kernel.mm.tuned_fused_int_mm_mul(mat1, mat2, mat3, out_dtype)
-
-
-def is_index_put_and_requires_h2d_sync_for_gpu_value(node):
-    from torch.fx.operator_schemas import normalize_function
-
-    if node.target not in [
-        torch.ops.aten.index_put.default,
-        torch.ops.aten.index_put_.default,
-    ]:
-        return False
-    # Inductor falls back to aten.index_put_.
-    # index_put_ will will call nonzero() and perform a H2D sync if
-    # any of its indices are bool/byte tensors
-    # However, it will short-circuit this H2D sync and run mask_fill_
-    # if the value we are putting is a cpu scalar.
-    # Therefore, when inductor sees an index_put_ with byte tensor indices,
-    # it should *not* convert the cpu scalar value into a gpu tensor.
-    args_, kwargs_ = normalize_function(node.target, node.args, node.kwargs)  # type: ignore[misc]
-    any_byte_bool_indices = False
-    indices = args_[1]
-    for i in indices:
-        if i is not None and i.meta["val"].dtype in [torch.bool, torch.int8]:
-            any_byte_bool_indices = True
-
-    val = args_[2].meta["val"]
-    val_is_cpu_scalar = val.device.type == "cpu" and val.numel() == 1
-    # If both these conditions hold, then converting the val
-    # to a gpu tensor will incur a H2D sync when inductor calls aten.index_put_
-    return any_byte_bool_indices and val_is_cpu_scalar
-
-
-class ConstructorMoverPass:
-    def __init__(self, target: str, allow_outputs: bool = False) -> None:
-        """
-        Move constructors from cpu to the target_device.
-
-        Sweeps through the module, looking for constructor nodes that can be moved
-        to the target_device.
-
-        A constructor node can be moved to the target_device iff all of its users
-        can also be moved (tested by cannot_be_moved). Otherwise, all dependent
-        constructor nodes won't be moved.
-
-        - target: target device type
-        - allow_outputs: allow outputs to be moved
-        """
-
-        self.target = target
-        self.allow_outputs = allow_outputs
-
-        assert isinstance(target, str), (
-            "target should be a string representing the device type. "
-            f"Got: {type(target).__name__}"
-        )
-
-    def allow_cpu_device(self, node: fx.Node) -> bool:
-        """
-        Returns whether a node that returns a tensor on the target device may have
-        cpu tensors as input.
-        """
-        return node.target in (
-            torch.ops.aten.index.Tensor,
-            torch.ops.aten.index_put.default,
-            torch.ops.aten.index_put_.default,
-            torch.ops.aten.copy.default,
-            torch.ops.aten.copy_.default,
-            torch.ops.aten.slice_scatter.default,
-        )
-
-    def cannot_be_moved(self, node: fx.Node) -> bool:
-        """
-        Returns whether a node can be moved to the target device.
-
-        If this function returns False, it means that this node and all of its users
-        won't be moved into the target device.
-        """
-        if node.target == "output":
-            return not self.allow_outputs
-
-        if not (
-            isinstance(node.target, torch._ops.OpOverload)
-            and node.target.namespace in ("prims", "aten")
-        ):
-            return True
-        if is_index_put_and_requires_h2d_sync_for_gpu_value(node):
-            return True
-
-        return False
-
-    def get_node_device(self, node: fx.Node) -> Optional[torch.device]:
-        """
-        Get the device of a node.
-        """
-        ten = node.meta.get("val")
-        return None if not isinstance(ten, torch.Tensor) else ten.device
-
-    def get_cpu_indeg_count(self, graph: fx.Graph) -> Dict[fx.Node, int]:
-        """
-        Get the number of cpu inputs to a node
-        """
-        cpu_indeg: Dict[fx.Node, int] = Counter()
-
-        for node in graph.nodes:
-            cpu_count = 0
-
-            def add_cpu_inp(node):
-                nonlocal cpu_count
-                device = self.get_node_device(node)
-                cpu_count += device is not None and device.type == "cpu"
-
-            pytree.tree_map_only(fx.Node, add_cpu_inp, (node.args, node.kwargs))
-
-            if cpu_count:
-                cpu_indeg[node] = cpu_count
-
-        return cpu_indeg
-
-    def __call__(self, graph: fx.Graph) -> None:
-        target_devices = set()
-        constructors = []
-
-        for node in graph.nodes:
-            device = self.get_node_device(node)
-            if device and device.type == self.target:
-                target_devices.add(device)
-
-            if not (
-                isinstance(node.target, torch._ops.OpOverload)
-                and node.target.namespace in ("prims", "aten")
-            ):
-                continue
-
-            if not torch._subclasses.fake_tensor._is_tensor_constructor(node.target):
-                continue
-
-            if not node.kwargs.get("device") == torch.device("cpu"):
-                continue
-
-            constructors.append(node)
-
-        # not handling multiple target devices initially
-        if not constructors or len(target_devices) != 1:
-            return
-
-        movable_constructors = self.find_movable_constructors(graph, constructors)
-
-        for node in movable_constructors:
-            kwargs = node.kwargs.copy()
-            kwargs["device"] = next(iter(target_devices))
-            node.kwargs = kwargs
-
-    def find_movable_constructors(
-        self, graph: fx.Graph, constructors: List[fx.Node]
-    ) -> Set[fx.Node]:
-        """
-        Starting from the cpu constructors, iterate through the graph and test that all of their
-        downstream uses can safely be moved to cpu.
-        """
-        cpu_indeg: Dict[fx.Node, int] = self.get_cpu_indeg_count(graph)
-
-        # which constructors cannot be moved to gpu
-        cannot_move_to_gpu: Set[fx.Node] = set()
-
-        # For any node in the graph, which constructors does it have a dependency on
-        constructor_dependencies: Dict[fx.Node, Set[fx.Node]] = defaultdict(set)
-
-        # if a cpu node has a dependency on two different cpu constructors,
-        # then if either constructor cannot be moved to gpu, the other cannot as well.
-        # In this case any node with a dependency on one will have a dependency on the other
-        equal_constructor_sets: Dict[fx.Node, Set[fx.Node]] = {
-            c: {c} for c in constructors
-        }
-
-        def make_dependencies_equivalent(
-            set1: Set[fx.Node], set2: Set[fx.Node]
-        ) -> Set[fx.Node]:
-            # could use union find but not worth complexity here
-            set1.update(set2)
-            for obj in set1:
-                equal_constructor_sets[obj] = set1
-            return set1
-
-        queue: List[fx.Node] = list(constructors)
-
-        for c in queue:
-            constructor_dependencies[c].add(c)
-
-        while queue:
-            node = queue.pop()
-            dependencies = constructor_dependencies[node]
-
-            for user in node.users:
-                if self.cannot_be_moved(user):
-                    cannot_move_to_gpu.update(dependencies)
-                    break
-
-                # this node was used on a op which takes in multiple devices and output a gpu
-                # tensor. we can convert its cpu input to gpu without making further changes
-                node_device = self.get_node_device(user)
-                if (
-                    self.allow_cpu_device(user)
-                    and node_device
-                    and node_device.type == self.target
-                ):
-                    del cpu_indeg[user]
-                else:
-                    # otherwise, we should continue look at its downstream uses
-                    cpu_indeg[user] -= 1
-                    if cpu_indeg[user] == 0:
-                        del cpu_indeg[user]
-                        queue.append(user)
-
-                unioned_set = make_dependencies_equivalent(
-                    dependencies, constructor_dependencies[user]
-                )
-                constructor_dependencies[user] = unioned_set
-
-        for node in cpu_indeg:
-            if constructor_dependencies[node]:
-                cannot_move_to_gpu.update(constructor_dependencies[node])
-
-        all_cannot_move_to_gpu = cannot_move_to_gpu.copy()
-        for constructor in cannot_move_to_gpu:
-            all_cannot_move_to_gpu.update(equal_constructor_sets[constructor])
-
-        return set(constructors) - all_cannot_move_to_gpu
-
-
-def move_constructors_to_gpu(graph: fx.Graph) -> None:
-    """
-    Moves intermediary tensors which are constructed on the cpu to gpu when safe
-    """
-    ConstructorMoverPass(get_gpu_type())(graph)
